Moves event into `/modules`
