+++ src/entities/calendarEvent/calendarEvent.model.ts
@@ -10,7 +10,7 @@ import { Connected, ConnectionType, withCursor } from '@utils/Connection'
 
 // Entities
 import { Subject } from '@entities/subject/subject.model'
-import { Event, ConnectedEvents } from '@entities/event/event.model'
+import { Event, ConnectedEvents } from '@modules/event/event.model'
 import { User, ConnectedUsers } from '@modules/user/user.model'
 import { VirtualLocation } from '../virtualLocation/virtualLocation.model'
