+++ src/entities/calendarEvent/calendarEvent.resolver.ts
@@ -28,8 +28,8 @@ import {
 
 // Services
 import { CalendarEventService } from './calendarEvent.service'
-import { EventService } from '../event/event.service'
-import { FollowerService } from '../follower/follower.service'
+import { EventService } from '@modules/event/event.service'
+import { FollowerService } from '@entities/follower/follower.service'
 @Resolver(() => CalendarEvent)
 export class CalendarEventResolver {
   constructor(
