+++ src/entities/follow/follow.resolver.ts
@@ -19,7 +19,7 @@ import { SchoolService } from '@modules/school/school.service'
 import { SubjectService } from '@entities/subject/subject.service'
 import { InstituteService } from '@entities/institute/institute.service'
 import { CalendarEventService } from '@entities/calendarEvent/calendarEvent.service'
-import { EventService } from '@entities/event/event.service'
+import { EventService } from '@modules/event/event.service'
 import { UserService } from '@modules/user/user.service'
 
 @Resolver(() => Follow)
