+++ src/entities/following/following.model.ts
@@ -12,7 +12,7 @@ import { School } from '@modules/school/school.model'
 import { Subject } from '@entities/subject/subject.model'
 import { Institute } from '@entities/institute/institute.model'
 import { CalendarEvent } from '@entities/calendarEvent/calendarEvent.model'
-import { Event } from '@entities/event/event.model'
+import { Event } from '@modules/event/event.model'
 import { User } from '@modules/user/user.model'
 
 const Resource = GQL.createUnionType({
