+++ src/modules/casl/casl-ability.factory.ts
@@ -12,7 +12,7 @@ import { Auth } from '@modules/auth/auth.model'
 import { CalendarEvent } from '@entities/calendarEvent/calendarEvent.model'
 import { Career } from '@entities/career/career.model'
 import { Discussion } from '@entities/discussion/discussion.model'
-import { Event } from '@entities/event/event.model'
+import { Event } from '@modules/event/event.model'
 import { Institute } from '@entities/institute/institute.model'
 import { School } from '@modules/school/school.model'
 import { Subject } from '@entities/subject/subject.model'
