+++ src/modules/event/event.inputs.ts
@@ -1,8 +1,8 @@
 import * as GQL from '@nestjs/graphql' // { Field, ObjectType, ID }
 import mongoose from 'mongoose'
 
-import { VirtualLocationInput } from '../virtualLocation/virtualLocation.inputs'
-import { CommentInput } from '../comment/comment.inputs'
+import { VirtualLocationInput } from '@entities/virtualLocation/virtualLocation.inputs'
+import { CommentInput } from '@entities/comment/comment.inputs'
 @GQL.InputType()
 export class CreateEventInput {
   @GQL.Field(() => String, { nullable: true })
